proxy/proxyports: move test-only code and use datapath FakeIptablesManager
Currently proxyPortsForTest is defined in testutils.go which will also
be compiled for non-test builds (due to the lack of a _test.go file
suffix), although it's only used in test code. Move it to
proxyports_test.go along with the other test code.

Also replace MockIPTablesManager by the same fake/mock implementation
already defined in pkg/datapath/fake/types.FakeIptablesManager.
