+++ pkg/proxy/proxyports/proxyports_test.go
@@ -8,12 +8,36 @@ import (
 	"os"
 	"testing"
 
+	"github.com/cilium/hive/hivetest"
 	"github.com/stretchr/testify/require"
 
+	datapath "github.com/cilium/cilium/pkg/datapath/fake/types"
 	"github.com/cilium/cilium/pkg/envoy"
 	"github.com/cilium/cilium/pkg/proxy/types"
 	"github.com/cilium/cilium/pkg/time"
+	"github.com/cilium/cilium/pkg/trigger"
 )
+
+func proxyPortsForTest(t *testing.T) (*ProxyPorts, func()) {
+	fakeIPTablesManager := &datapath.FakeIptablesManager{}
+	config := ProxyPortsConfig{
+		ProxyPortrangeMin:          10000,
+		ProxyPortrangeMax:          20000,
+		RestoredProxyPortsAgeLimit: 0,
+	}
+
+	p := NewProxyPorts(hivetest.Logger(t), config, fakeIPTablesManager)
+	triggerDone := make(chan struct{})
+	p.Trigger, _ = trigger.NewTrigger(trigger.Parameters{
+		MinInterval:  10 * time.Millisecond,
+		TriggerFunc:  func(reasons []string) {},
+		ShutdownFunc: func() { close(triggerDone) },
+	})
+	return p, func() {
+		p.Trigger.Shutdown()
+		<-triggerDone
+	}
+}
 
 func (p *ProxyPorts) released(pp *ProxyPort) bool {
 	p.mutex.Lock()
